refactoring: created news classes to Builder, Manager Execute and refactored functions to better code readability
